requirements.txt
pydantic>=1.10.0
pydantic-settings==2.10.1
email-validator>=2.0.0
loguru>=0.7.0
psutil>=5.9.0
faker>=18.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
beartype>=0.18.0
langchain-openai
fastapi
redis
uvicorn>=0.21.0
gunicorn>=21.2.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.8
alembic>=1.10.0
# psycopg2-binary>=2.9.0  # REMOVED - Using async-only pattern with asyncpg
asyncpg>=0.29.0
cloud-sql-python-connector[asyncpg]>=1.7.0
aiosqlite>=0.19.0  # For testing only
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
argon2-cffi>=23.1.0
python-multipart>=0.0.5
google-cloud-secret-manager>=2.16.0
authlib>=1.0.0
python-dateutil>=2.8.2
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.0.0
httpx>=0.23.0
clickhouse-connect
clickhouse-driver
cryptography>=3.4.0
langgraph
starlette
itsdangerous
pytest
langfuse
google-cloud-secret-manager
langchain-google-genai 
pydantic-to-typescript
tenacity
rich
pytest-asyncio
langchain-community
pytest-mock
watchfiles
fastmcp>=0.5.0
mcp>=1.0.0
google-cloud-error-reporting>=1.9.0